Use size_t where it is appropriate.
In src/PrecedenceGraph.h:

>      /** Get all successor SCCs of a given SCC. */
-    const std::set<unsigned>& getSuccessorSCCs(const unsigned scc) const {
+    const std::set<int>& getSuccessorSCCs(const int scc) const {

successors in successors.at(scc) is a vector, so scc is the index. size_t seems more appropriate. Similarly for the other getX(scc) methods. It's an identifier, but it's an index used as an identifier.

In src/PrecedenceGraph.h:

> @@ -110,17 +110,17 @@ class SCCGraph : public AstAnalysis {
     std::map<const AstRelation*, int> relationToScc;

     /** Adjacency lists for the SCC graph */
-    std::vector<std::set<unsigned>> successors;
+    std::vector<std::set<int>> successors;

Aren't these sets of indices into containers, so unsigned values are more appropriate?