added c++ interface to purge relations
The c++ interface for Relation has a method purge() that eliminates
all the facts in the relation.  Additionally, SouffleProgram has the
methods purgeInternalRelations() an purgeOutputRelations() that purge
all the relations in those categories.  Finally, a small change has
been made in the Synthesiser so all non-temorary relations appear in
the list of relations of the souffleProgram object.